more reshuffling for tests
