+++ packages/lambda/src/api/clean-items.ts
@@ -1,6 +1,6 @@
 import {DeleteObjectCommand} from '@aws-sdk/client-s3';
 import type {AwsRegion} from '../regions';
-import {getS3Client} from '../shared/aws-clients';
+import {getS3Client} from '../shared/get-s3-client';
 import {pLimit} from '../shared/p-limit';
 
 const limit = pLimit(10);
