+++ packages/lambda/src/api/create-bucket.ts
@@ -6,7 +6,7 @@ import {
 } from '@aws-sdk/client-s3';
 import type {ProviderSpecifics} from '@remotion/serverless';
 import type {AwsRegion} from '../regions';
-import {getS3Client} from '../shared/aws-clients';
+import {getS3Client} from '../shared/get-s3-client';
 
 export const createBucket: ProviderSpecifics<AwsRegion>['createBucket'] =
 	async ({region, bucketName}) => {
