+++ packages/lambda/src/shared/aws-clients.ts
@@ -1,245 +1,10 @@
-import {CloudWatchLogsClient} from '@aws-sdk/client-cloudwatch-logs';
-import {IAMClient} from '@aws-sdk/client-iam';
-import {LambdaClient} from '@aws-sdk/client-lambda';
-import {S3Client} from '@aws-sdk/client-s3';
-import {ServiceQuotasClient} from '@aws-sdk/client-service-quotas';
-import {STSClient} from '@aws-sdk/client-sts';
-import {fromIni} from '@aws-sdk/credential-providers';
-import type {CustomCredentials} from '@remotion/serverless/client';
-import {random} from 'remotion/no-react';
+import type {CloudWatchLogsClient} from '@aws-sdk/client-cloudwatch-logs';
+import type {IAMClient} from '@aws-sdk/client-iam';
+import type {LambdaClient} from '@aws-sdk/client-lambda';
+import type {ServiceQuotasClient} from '@aws-sdk/client-service-quotas';
+import type {STSClient} from '@aws-sdk/client-sts';
 import type {AwsRegion} from '../regions';
-import {checkCredentials} from './check-credentials';
-import {MAX_FUNCTIONS_PER_RENDER} from './constants';
-import {isInsideLambda} from './is-in-lambda';
-
-const _clients: Partial<
-	Record<
-		string,
-		| CloudWatchLogsClient
-		| LambdaClient
-		| S3Client
-		| IAMClient
-		| ServiceQuotasClient
-		| STSClient
-	>
-> = {};
-
-type CredentialPair = {
-	accessKeyId: string;
-	secretAccessKey: string;
-	sessionToken?: string;
-};
-type AwsCredentialIdentityProvider = ReturnType<typeof fromIni>;
-
-const getCredentials = ():
-	| CredentialPair
-	| AwsCredentialIdentityProvider
-	| undefined => {
-	if (isInsideLambda()) {
-		return undefined;
-	}
-
-	if (process.env.REMOTION_AWS_PROFILE) {
-		return fromIni({
-			profile: process.env.REMOTION_AWS_PROFILE,
-		});
-	}
-
-	if (
-		process.env.REMOTION_AWS_ACCESS_KEY_ID &&
-		process.env.REMOTION_AWS_SECRET_ACCESS_KEY &&
-		process.env.REMOTION_AWS_SESSION_TOKEN
-	) {
-		return {
-			accessKeyId: process.env.REMOTION_AWS_ACCESS_KEY_ID,
-			secretAccessKey: process.env.REMOTION_AWS_SECRET_ACCESS_KEY,
-			sessionToken: process.env.REMOTION_AWS_SESSION_TOKEN,
-		};
-	}
-
-	if (
-		process.env.REMOTION_AWS_ACCESS_KEY_ID &&
-		process.env.REMOTION_AWS_SECRET_ACCESS_KEY
-	) {
-		return {
-			accessKeyId: process.env.REMOTION_AWS_ACCESS_KEY_ID,
-			secretAccessKey: process.env.REMOTION_AWS_SECRET_ACCESS_KEY,
-		};
-	}
-
-	if (process.env.AWS_PROFILE) {
-		return fromIni({
-			profile: process.env.AWS_PROFILE,
-		});
-	}
-
-	if (
-		process.env.AWS_ACCESS_KEY_ID &&
-		process.env.AWS_SECRET_ACCESS_KEY &&
-		process.env.AWS_SESSION_TOKEN
-	) {
-		return {
-			accessKeyId: process.env.AWS_ACCESS_KEY_ID as string,
-			secretAccessKey: process.env.AWS_SECRET_ACCESS_KEY as string,
-			sessionToken: process.env.AWS_SESSION_TOKEN as string,
-		};
-	}
-
-	if (process.env.AWS_ACCESS_KEY_ID && process.env.AWS_SECRET_ACCESS_KEY) {
-		return {
-			accessKeyId: process.env.AWS_ACCESS_KEY_ID as string,
-			secretAccessKey: process.env.AWS_SECRET_ACCESS_KEY as string,
-		};
-	}
-
-	return undefined;
-};
-
-const getCredentialsHash = ({
-	customCredentials,
-	region,
-	service,
-}: {
-	region: AwsRegion;
-	customCredentials: CustomCredentials<AwsRegion> | null;
-	service: keyof ServiceMapping;
-}): string => {
-	const hashComponents: {[key: string]: unknown} = {};
-
-	if (process.env.REMOTION_SKIP_AWS_CREDENTIALS_CHECK) {
-		hashComponents.credentials = {
-			credentialsSkipped: true,
-		};
-	} else if (process.env.REMOTION_AWS_PROFILE) {
-		hashComponents.credentials = {
-			awsProfile: process.env.REMOTION_AWS_PROFILE,
-		};
-	} else if (
-		process.env.REMOTION_AWS_ACCESS_KEY_ID &&
-		process.env.REMOTION_AWS_SECRET_ACCESS_KEY
-	) {
-		hashComponents.credentials = {
-			accessKeyId: process.env.REMOTION_AWS_ACCESS_KEY_ID,
-			secretAccessKey: process.env.REMOTION_AWS_SECRET_ACCESS_KEY,
-		};
-	} else if (process.env.AWS_PROFILE) {
-		hashComponents.credentials = {
-			awsProfile: process.env.AWS_PROFILE,
-		};
-	} else if (
-		process.env.AWS_ACCESS_KEY_ID &&
-		process.env.AWS_SECRET_ACCESS_KEY
-	) {
-		hashComponents.credentials = {
-			accessKeyId: process.env.AWS_ACCESS_KEY_ID as string,
-			secretAccessKey: process.env.AWS_SECRET_ACCESS_KEY as string,
-		};
-	}
-
-	hashComponents.customCredentials = customCredentials;
-	hashComponents.region = region;
-	hashComponents.service = service;
-
-	return random(JSON.stringify(hashComponents)).toString().replace('0.', '');
-};
-
-export type ServiceMapping = {
-	s3: S3Client;
-	cloudwatch: CloudWatchLogsClient;
-	iam: IAMClient;
-	lambda: LambdaClient;
-	servicequotas: ServiceQuotasClient;
-	sts: STSClient;
-};
-
-export const getServiceClient = <T extends keyof ServiceMapping>({
-	region,
-	service,
-	customCredentials,
-}: {
-	region: AwsRegion;
-	service: T;
-	customCredentials: CustomCredentials<AwsRegion> | null;
-}): ServiceMapping[T] => {
-	const Client = (() => {
-		if (service === 'cloudwatch') {
-			return CloudWatchLogsClient;
-		}
-
-		if (service === 'lambda') {
-			return LambdaClient;
-		}
-
-		if (service === 's3') {
-			return S3Client;
-		}
-
-		if (service === 'iam') {
-			return IAMClient;
-		}
-
-		if (service === 'servicequotas') {
-			return ServiceQuotasClient;
-		}
-
-		if (service === 'sts') {
-			return STSClient;
-		}
-
-		throw new TypeError('unknown client ' + service);
-	})();
-
-	const key = getCredentialsHash({
-		region,
-		customCredentials,
-		service,
-	});
-
-	if (!_clients[key]) {
-		checkCredentials();
-
-		const lambdaOptions =
-			service === 'lambda'
-				? {
-						httpsAgent: {
-							maxSockets: MAX_FUNCTIONS_PER_RENDER * 2,
-						},
-					}
-				: undefined;
-
-		const client = customCredentials
-			? new Client({
-					region: customCredentials.region ?? 'us-east-1',
-					credentials:
-						customCredentials.accessKeyId && customCredentials.secretAccessKey
-							? {
-									accessKeyId: customCredentials.accessKeyId,
-									secretAccessKey: customCredentials.secretAccessKey,
-								}
-							: undefined,
-					endpoint: customCredentials.endpoint,
-					requestHandler: lambdaOptions,
-				})
-			: process.env.REMOTION_SKIP_AWS_CREDENTIALS_CHECK
-				? new Client({
-						region,
-						requestHandler: lambdaOptions,
-					})
-				: new Client({
-						region,
-						credentials: getCredentials(),
-						requestHandler: lambdaOptions,
-					});
-
-		if (process.env.REMOTION_DISABLE_AWS_CLIENT_CACHE) {
-			return client as ServiceMapping[T];
-		}
-
-		_clients[key] = client;
-	}
-
-	return _clients[key] as ServiceMapping[T];
-};
+import {getServiceClient} from './get-service-client';
 
 export const getCloudWatchLogsClient = (
 	region: AwsRegion,
@@ -248,17 +13,6 @@ export const getCloudWatchLogsClient = (
 		region,
 		service: 'cloudwatch',
 		customCredentials: null,
-	});
-};
-
-export const getS3Client = (
-	region: AwsRegion,
-	customCredentials: CustomCredentials<AwsRegion> | null,
-): S3Client => {
-	return getServiceClient({
-		region: customCredentials?.region ?? region,
-		service: 's3',
-		customCredentials,
 	});
 };
