+++ packages/lambda/src/test/integration/renders/muted-render.test.ts
@@ -5,7 +5,7 @@ import path from 'path';
 import {afterAll, expect, test} from 'vitest';
 import {deleteRender} from '../../../api/delete-render';
 import {rendersPrefix} from '../../../defaults';
-import {lambdaLs} from '../../../functions/helpers/io';
+import {mockImplementation} from '../../mock-implementation';
 import {simulateLambdaRender} from '../simulate-lambda-render';
 
 afterAll(async () => {
@@ -47,7 +47,7 @@ test('Should make muted render audio', async () => {
 
 	unlinkSync(tmpfile);
 
-	const files = await lambdaLs({
+	const files = await mockImplementation.listObjects({
 		bucketName: progress.outBucket as string,
 		region: 'eu-central-1',
 		expectedBucketOwner: 'abc',
@@ -62,7 +62,7 @@ test('Should make muted render audio', async () => {
 		renderId,
 	});
 
-	const expectFiles = await lambdaLs({
+	const expectFiles = await mockImplementation.listObjects({
 		bucketName: progress.outBucket as string,
 		region: 'eu-central-1',
 		expectedBucketOwner: 'abc',
