+++ packages/lambda/src/test/integration/renders/transparent-webm.test.ts
@@ -5,7 +5,7 @@ import path from 'node:path';
 import {afterAll, expect, test} from 'vitest';
 import {deleteRender} from '../../../api/delete-render';
 import {rendersPrefix} from '../../../defaults';
-import {lambdaLs} from '../../../functions/helpers/io';
+import {mockImplementation} from '../../mock-implementation';
 import {simulateLambdaRender} from '../simulate-lambda-render';
 
 afterAll(async () => {
@@ -46,7 +46,7 @@ test('Should make a transparent video', async () => {
 	expect(probe.stderr).toMatch(/Audio: opus, 48000 Hz/);
 	fs.unlinkSync(out);
 
-	const files = await lambdaLs({
+	const files = await mockImplementation.listObjects({
 		bucketName: progress.outBucket as string,
 		region: 'eu-central-1',
 		expectedBucketOwner: 'abc',
@@ -61,7 +61,7 @@ test('Should make a transparent video', async () => {
 		renderId,
 	});
 
-	const expectFiles = await lambdaLs({
+	const expectFiles = await mockImplementation.listObjects({
 		bucketName: progress.outBucket as string,
 		region: 'eu-central-1',
 		expectedBucketOwner: 'abc',
