+++ packages/lambda/src/test/mock-implementation.ts
@@ -1,6 +1,10 @@
 import type {ProviderSpecifics} from '@remotion/serverless';
 import type {AwsRegion} from '../regions';
-import {addMockBucket, getMockBuckets} from './mocks/mock-store';
+import {
+	addMockBucket,
+	getMockBuckets,
+	getS3FilesInBucket,
+} from './mocks/mock-store';
 
 export const mockImplementation: ProviderSpecifics<AwsRegion> = {
 	applyLifeCycle: () => Promise.resolve(),
@@ -18,4 +22,31 @@ export const mockImplementation: ProviderSpecifics<AwsRegion> = {
 		return Promise.resolve();
 	},
 	getBuckets: () => Promise.resolve(getMockBuckets()),
+	listObjects: (input) => {
+		if (!input) {
+			throw new Error('need to pass input');
+		}
+
+		const files = getS3FilesInBucket({
+			bucketName: input.bucketName,
+			region: input.region,
+		});
+
+		return Promise.resolve(
+			files
+				.filter((p) => p.key.startsWith(input.prefix))
+				.map((file) => {
+					const size =
+						typeof file.content === 'string' ? file.content.length : 0;
+					return {
+						Key: file.key,
+						ETag: 'etag',
+						LastModified: new Date(0),
+						Owner: undefined,
+						Size: size,
+						StorageClass: undefined,
+					};
+				}),
+		);
+	},
 };
